Add files via upload
We made 2 versions of this app to conduct an A/B testing. The version B showcased more positive results as a result of which I am uploading only the "B" version of the prototype.
It can be simulated on your local machine using JustInMind tool. The tool can be downloaded here: http://www.justinmind.com/download